+++ src/Zen.SpectreConsole.Extensions/ISpectreConfiguration.cs
@@ -1,0 +1,9 @@
+using Spectre.Console.Cli;
+
+namespace Zen.SpectreConsole.Extensions
+{
+    public interface ISpectreConfiguration
+    {
+        void ConfigureCommandApp(in IConfigurator configurator);
+    }
+}